backend/requirements.txt
pyperclip
youtube_transcript_api
pytesseract
googletrans==3.1.0a0
flask==2.2.2
numpy==1.23.3
pandas==1.5.0
opencv-python==4.6.0.66 
underthesea==1.3.5a3  
pyvi==0.1.1 
translate==3.6.1 
faiss-cpu
ftfy==6.1.1 
regex==2022.9.13
tqdm==4.64.1
matplotlib==3.6.0
langdetect==1.0.9
rapidfuzz==2.12.0
sentence_transformers==2.2.2
werkzeug==2.2.2